added error check for ARMCL build

+++ package/lib-armcl-cpu-master/build.sh
@@ -20,8 +20,15 @@ cd ${INSTALL_DIR}/src/obj
 
 ${CK_MAKE} -j ${CK_HOST_CPU_NUMBER_OF_PROCESSORS} -f ${ORIGINAL_PACKAGE_DIR}/Makefile $@
 
+if [ "${?}" != "0" ] ; then
+  echo "Error: ARMCL build failed ..."
+  exit 1
+fi
+
 cd ..
 rm -rf ../install/include
 mkdir ../install/include
 cp -rf arm_compute ../install/include
 cp -rf tests ../install/include/arm_compute
+
+exit 0
